Move watsonhelpers.py to watson module
