+++ qwatson/watson/watsonhelpers.py
